requirements.txt
streamlit>=1.28.0
nltk>=3.8.1
spacy>=3.7.0
scikit-learn>=1.3.0
pymongo>=4.6.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.1.0
plotly>=5.18.0
requests>=2.31.0

